+++ src/api/resources/transactions/types/GetEarnedRewardsResponse.ts
@@ -6,27 +6,13 @@ import type * as KardApi from "../../../index.js";
  * @example
  *     {
  *         data: [{
- *                 type: "rewardedTransaction",
+ *                 type: "approvedTransaction",
  *                 id: "fcabf024-3870-41f3-9fbd-b43ea85a3d19",
  *                 attributes: {
- *                     status: KardApi.RewardedTransactionStatus.Approved,
+ *                     status: "APPROVED",
  *                     transactionId: "TXN-20241001-F21-127964",
  *                     transactionAmountInCents: 12796,
- *                     transactionTimestamp: "2024-10-01T01:36:57Z",
- *                     paidToIssuer: KardApi.PaymentStatus.Unpaid,
- *                     payoutTimestamp: "2024-10-01T02:00:00Z",
- *                     cardBIN: "123456",
- *                     cardLastFour: "4321",
- *                     commissionEarned: {
- *                         issuer: {
- *                             type: KardApi.CommissionValueType.Cents,
- *                             value: 102
- *                         },
- *                         user: {
- *                             type: KardApi.CommissionValueType.Cents,
- *                             value: 320
- *                         }
- *                     }
+ *                     transactionTimestamp: "2024-10-01T01:36:57Z"
  *                 },
  *                 relationships: {
  *                     user: {
@@ -52,19 +38,13 @@ import type * as KardApi from "../../../index.js";
  *                 type: "rewardedTransaction",
  *                 id: "7bcbdb95-f3a5-4f56-a9be-4c25f313eb0a",
  *                 attributes: {
- *                     status: KardApi.RewardedTransactionStatus.Settled,
+ *                     status: "SETTLED",
  *                     transactionId: "TXN-20240928-TGT-778813",
  *                     transactionAmountInCents: 8800,
  *                     transactionTimestamp: "2024-09-28T14:11:22Z",
  *                     paidToIssuer: KardApi.PaymentStatus.PaidInFull,
  *                     payoutTimestamp: "2024-09-29T10:15:00Z",
- *                     cardBIN: "123456",
- *                     cardLastFour: "4321",
  *                     commissionEarned: {
- *                         issuer: {
- *                             type: KardApi.CommissionValueType.Cents,
- *                             value: 70
- *                         },
  *                         user: {
  *                             type: KardApi.CommissionValueType.Cents,
  *                             value: 220